Reordering Staff Commands
In the event of creating new ranks and adjusting the hierarchy:
I have made multiple changes to places that have .isAtLeast() .isStaff() and the Enum Rank to reorder the staff team.

The Ranks from lowest to highest in sort of privs will be:
Builder
Dev
Trial
Mod
HR
Admin

The Vanilla OP has been adjusted to MOD rank and Up
While the IsStaff() Command has been adjusted to include Builders and Up.